backend/paddle/requirements.txt
flask==2.*
flask-babel==3.*
flask-cors==3.*
Flask-SocketIO==5.*
Flask-Session==0.4.*
python-engineio==4.3.3
python-socketio==5.7.1
eventlet==0.30.2
gevent==22.10.2
gunicorn==20.1.0
geventhttpclient==2.0.2
opencv-contrib-python-headless==4.6.0.66
opencv-python-headless==4.6.0.66
pdf2image==1.16.3
pdf2docx==0.5.6
paddleocr==2.6.1.3
python-dotenv==0.21.1